sentinel_demo/requirements.txt
fastapi
uvicorn
jinja2
python-dateutil==2.9.0.post0
psutil
requests
